Fix alpha bleed algorithm (#93)
This PR makes a couple of fixes / adjustments to the alpha bleed
algorithm.

1. It only bleeds one pixel now. See this
[README](https://github.com/EgoMoose/alpha-bleed/blob/main/README.md)
for why.
2. Fully transparent pixels that are not bled into are converted to pure
black for better encoding.
3. Fixes a mistake in the algorithm where pixels could sample neighbors
from the same queue cycle.

The first pass iterates over pixels left-to-right, top-to-bottom so this
effect naturally impacts pixels that come further along in the iteration
(right and bottom). However, due to the order of the neighbor offsets
this effect is somewhat muddied. Regardless, pixels should not be
sampling their bled neighbors from the same cycle.

Given an input image:

![tree](https://github.com/user-attachments/assets/05152ece-c1f6-47a8-925b-25c019adba6a)

The results go from looking like this (everything fully opaque):

![image](https://github.com/user-attachments/assets/f6f7df8e-c2a5-45b3-8895-51f70c5e2142)

To this:

![image](https://github.com/user-attachments/assets/efd832ac-c90b-4760-a795-d73776455ba1)