Bleed entire image (#94)
Hi again, very sorry I didn't get this in with the last PR, but I quite
literally just figured out this was an issue independantly 10ish minutes
ago.

Long story short I was wrong about only needing to bleed one pixel. This
is only applicable when the image is being upscaled, not when it's being
down-scaled.

If you want more info on how / why I came to this conclusion you can
read the last couple responses in this devforum post.
https://devforum.roblox.com/t/photobooth-plugin/3401720/80